Introduced legacy-reducer. Moved reducers to directory and files where they belong.
