+++ src/app/App.jsx
@@ -15,7 +15,7 @@ import { isMapSubPage } from './routes';
 import CatalogDetailContainer from './containers/CatalogDetailContainer';
 import CatalogSearchContainer from './containers/CatalogSearchContainer';
 import QuerySearchContainer from './containers/QuerySearchContainer';
-import { getCurrentPage } from '../reducers/current-page-reducer';
+import { getCurrentPage } from '../shared/ducks/current-page/current-page-reducer';
 import { isEmbedded, isEmbedPreview, isPrintMode, isPrintModeLandscape, isInPrintorEmbedMode } from '../shared/ducks/ui/ui';
 import { CMS_PAGE_MAPPING } from './pages/CMSPageMapping';
