+++ src/header/containers/header-search/HeaderSearchContainer.test.jsx
@@ -7,14 +7,14 @@ import { getSuggestions, setActiveSuggestion, FETCH_SUGGESTIONS_REQUEST, SET_ACT
 
 import { fetchDataSelection, fetchSearchResultsByQuery, FETCH_DATA_SELECTION, FETCH_SEARCH_RESULTS_BY_QUERY } from '../../ducks/search/search';
 
-import { fetchDetail, FETCH_DETAIL } from '../../../reducers/details';
 import piwikTracker from '../../../shared/services/piwik-tracker/piwik-tracker';
+import { fetchDetail, FETCH_DETAIL } from '../../../shared/ducks/detail/detail';
 import { routing } from '../../../app/routes';
 import PAGES from '../../../app/pages';
 
 jest.mock('../../ducks/auto-suggest/auto-suggest');
-jest.mock('../../../reducers/details');
 jest.mock('../../../shared/services/piwik-tracker/piwik-tracker');
+jest.mock('../../../shared/ducks/detail/detail');
 jest.mock('../../ducks/search/search');
 
 describe('HeaderSearchContainer', () => {
