+++ src/map/containers/map/MapContainer.integration.test.jsx
@@ -12,10 +12,10 @@ import MapDetailReducer from '../../ducks/detail/map-detail';
 import MapLayersReducer from '../../ducks/layers/map-layers';
 import MapReducer from '../../ducks/map/map';
 import PanelLayersReducer from '../../ducks/panel-layers/map-panel-layers';
-import MapSearchResultsReducer from '../../ducks/search-results/map-search-results';
+import MapSearchResultsReducer from '../../../shared/ducks/search/search';
 
 import UiReducer from '../../../shared/ducks/ui/ui';
-import UserReducer from '../../../reducers/user';
+import UserReducer from '../../../shared/ducks/user/user';
 
 const mapLayersReducer = combineReducers({
   layers: MapLayersReducer,
