+++ src/map/containers/preview-panel/MapPreviewPanel.jsx
@@ -11,7 +11,7 @@ import MapDetailResult from '../../components/detail-result/MapDetailResult';
 import MapSearchResults from '../../components/search-results/MapSearchResults';
 import { getMapDetail } from '../../ducks/detail/map-detail';
 import { getPanoPreview } from '../../../pano/ducks/preview/pano-preview';
-import { getMapSearchResults } from '../../ducks/search-results/map-search-results';
+import { getMapSearchResults } from '../../../shared/ducks/search/search';
 
 const previewPanelSearchResultLimit = 3;
