+++ src/map/ducks/preview-panel/map-preview-panel.js
@@ -1,4 +1,4 @@
-import { FETCH_SEARCH_RESULTS_BY_LOCATION } from '../../../shared/actions';
+import { FETCH_SEARCH_RESULTS_BY_LOCATION } from '../../../shared/ducks/search/search';
 import { UPDATE_MAP } from '../map/map';
 import { routing } from '../../../app/routes';
