+++ src/map/ducks/preview-panel/map-preview-panel.test.js
@@ -6,7 +6,7 @@ import reducer, {
   showDetailView,
   showSearchView
 } from './map-preview-panel';
-import { FETCH_SEARCH_RESULTS_BY_LOCATION } from '../../../shared/actions';
+import { FETCH_SEARCH_RESULTS_BY_LOCATION } from '../../../shared/ducks/search/search';
 
 const initialState = {};
 describe('map preview reducer and actions', () => {
