+++ src/shared/actions.js
@@ -7,11 +7,9 @@ import {
 
 export const REQUEST_GEOSEARCH = 'REQUEST_GEOSEARCH';
 export const REQUEST_NEAREST_DETAILS = 'REQUEST_NEAREST_DETAILS';
-export const FETCH_SEARCH_RESULTS_BY_LOCATION = 'FETCH_SEARCH_RESULTS_BY_LOCATION';
 
 export default {
   // change url
-  SHOW_SEARCH_RESULTS: 'SHOW_SEARCH_RESULTS',
   MAP_END_DRAWING: 'MAP_END_DRAWING',
   SHOW_DETAIL: 'SHOW_DETAIL',
   FETCH_STRAATBEELD_BY_ID: 'FETCH_STRAATBEELD_BY_ID',
@@ -24,9 +22,6 @@ export default {
 
   // ignore
   URL_CHANGE: 'URL_CHANGE',
-
-  FETCH_SEARCH_RESULTS_BY_QUERY: 'FETCH_SEARCH_RESULTS_BY_QUERY',
-  FETCH_SEARCH_RESULTS_BY_LOCATION,
   FETCH_SEARCH_RESULTS_CATEGORY: 'FETCH_SEARCH_RESULTS_CATEGORY',
   MAP_ADD_PANO_OVERLAY,
   MAP_REMOVE_PANO_OVERLAY,
