+++ src/shared/ducks/ui/ui.js
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { isMapCurrentPage } from '../../../reducers/current-page-reducer';
+import { isMapCurrentPage } from '../current-page/current-page-reducer';
 import { getStraatbeeld } from '../straatbeeld/straatbeeld';
 
 export const REDUCER_KEY = 'ui';
