+++ src/shared/ducks/user/user.test.js
