+++ src/store/store.js
@@ -6,7 +6,7 @@ import queryString from 'querystring';
 import rootSaga from '../root-saga';
 import '../shared/ducks/error-message';
 import * as auth from '../shared/services/auth/auth';
-import { authenticateUser } from '../reducers/user';
+import { authenticateUser } from '../shared/ducks/user/user';
 import { fetchCatalogFilters } from '../catalog/ducks/data-selection/data-selection-catalog';
 import rootReducer from '../reducers/root';
 import documentHeadMiddleware from './middleware/documentHead';
